# backend/internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: LLM Monitoring & Deci.Scoring API
  description: >-
    Records browser-run LLM interactions (WebLLM/Gemma) and grades them with a
    transparent, rule-based Deci.Scoring engine. Authentication uses JWT
    access tokens with rotating refresh tokens.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost:8080
    description: Local development
tags:
  - name: meta
    description: Public config, health and readiness
  - name: auth
    description: Registration, login, sessions and profile
  - name: llm
    description: LLM runs, decision scores and aggregate metrics

paths:
  /config:
    get:
      tags: [meta]
      summary: Public (non-secret) app configuration
      responses:
        "200": { description: Configuration object }
  /version:
    get:
      tags: [meta]
      summary: Build name and version
      responses:
        "200": { description: Version object }
  /health:
    get:
      tags: [meta]
      summary: Liveness probe
      responses:
        "200": { description: Process is up }
  /ready:
    get:
      tags: [meta]
      summary: Readiness probe (checks database)
      responses:
        "200": { description: Ready to serve }
        "500": { description: Database unavailable }

  /auth/register:
    post:
      tags: [auth]
      summary: Create an account and log in
      description: Rate limited by client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email and password
      description: Rate limited by client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate a refresh token for a new token pair
      description: >-
        The presented refresh token is revoked and a new pair issued. Presenting
        a token that was already revoked is treated as evidence of theft —
        rotation means the legitimate holder has since swapped it — so every
        session for that account is revoked and the caller must sign in again.

      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the presented refresh token
      description: >-
        Unauthenticated by design, so a client whose access token has already
        expired can still retire its refresh token. Rate limited by client IP.
        Always returns 200, whether or not the token matched, so it cannot be
        used to probe which tokens exist.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200": { description: Logged out }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/me:
    get:
      tags: [auth]
      summary: Current authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [auth]
      summary: Update the current user's profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change the current user's password
      description: >-
        Verifies the current password, sets the new one, and revokes every
        refresh token for the account — including the caller's, so a stolen
        token cannot outlive the change. A fresh token pair is returned in its
        place, so the caller stays signed in while every other device is signed
        out. Clients must replace their stored tokens with this response.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Password changed; all sessions revoked and a new token pair issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/sessions:
    get:
      tags: [auth]
      summary: List the current user's sessions
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
                  count: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/sessions/{id}:
    delete:
      tags: [auth]
      summary: Revoke one of the current user's sessions
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /llm/models:
    get:
      tags: [llm]
      summary: List browser-runnable models
      responses:
        "200":
          description: Models
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items: { $ref: "#/components/schemas/ModelInfo" }
  /llm/runs:
    post:
      tags: [llm]
      summary: Record a new LLM run (optionally auto-score it)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRunRequest" }
      responses:
        "201":
          description: Created run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      tags: [llm]
      summary: List the current user's runs (cursor-paginated)
      description: >-
        Returns run summaries newest-first. Paging is by cursor: pass the
        previous page's `next_cursor` as `before`. Summaries omit the prompt,
        response and system_prompt bodies — fetch GET /llm/runs/{id} for those.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - name: before
          in: query
          description: Return runs created strictly before this instant. Omit for the newest page.
          schema: { type: string, format: date-time }
        - { name: model, in: query, schema: { type: string } }
      responses:
        "200":
          description: A page of run summaries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /llm/runs/{id}:
    get:
      tags: [llm]
      summary: Get a single run
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [llm]
      summary: Delete a run
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /llm/runs/{id}/score:
    post:
      tags: [llm]
      summary: Score a run (optionally with custom weights)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScoreRequest" }
      responses:
        "200":
          description: Score
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Score" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    get:
      tags: [llm]
      summary: Get a run's existing score
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Score
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Score" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /llm/metrics:
    get:
      tags: [llm]
      summary: Aggregate dashboard metrics for the current user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Metrics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Metrics" }
        "401": { $ref: "#/components/responses/Unauthorized" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict (e.g. email already registered)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: Stable machine-readable code }
        message: { type: string }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: string }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 8 }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, description: Access token lifetime in seconds }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Session:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_agent: { type: string }
        ip_address: { type: string }
        expires_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    ModelInfo:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        family: { type: string }
        size_hint: { type: string }
        recommended: { type: boolean }
    Weights:
      type: object
      description: Scoring weights; components sum to 1.0.
      properties:
        completion: { type: number }
        latency: { type: number }
        efficiency: { type: number }
        keywords: { type: number }
        length: { type: number }
    CreateRunRequest:
      type: object
      required: [model, prompt, response]
      properties:
        model: { type: string }
        prompt: { type: string }
        response: { type: string }
        system_prompt: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        latency_ms: { type: integer }
        temperature: { type: number }
        expected_keywords:
          type: array
          items: { type: string }
        metadata: { type: object, additionalProperties: true }
        auto_score: { type: boolean, description: Score immediately on create }
    ScoreRequest:
      type: object
      properties:
        weights: { $ref: "#/components/schemas/Weights" }
    Score:
      type: object
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        score: { type: number, description: 0..100 }
        grade: { type: string, enum: [A, B, C, D, F] }
        breakdown: { $ref: "#/components/schemas/Breakdown" }
        rationale: { type: string }
        created_at: { type: string, format: date-time }
    Breakdown:
      type: object
      description: Per-component scores, each 0..100.
      properties:
        completion: { type: number }
        latency: { type: number }
        efficiency: { type: number }
        keywords: { type: number }
        length: { type: number }
    Run:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        model: { type: string }
        prompt: { type: string }
        response: { type: string }
        system_prompt: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        latency_ms: { type: integer }
        temperature: { type: number }
        expected_keywords:
          type: array
          items: { type: string }
        metadata: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        score: { $ref: "#/components/schemas/Score" }
    Metrics:
      type: object
      properties:
        total_runs: { type: integer }
        scored_runs: { type: integer }
        avg_score: { type: number }
        avg_latency_ms: { type: number }
        avg_completion_tokens: { type: number }
        runs_by_model:
          type: object
          additionalProperties: { type: integer }
        grade_distribution:
          type: object
          additionalProperties: { type: integer }
    RunSummary:
      type: object
      description: >-
        List-view projection of a Run. The large text columns are intentionally
        absent so listing a page does not transfer data the list never renders.
      properties:
        id: { type: string, format: uuid }
        model: { type: string }
        prompt_preview: { type: string, description: First 160 characters of the prompt. }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        latency_ms: { type: integer }
        created_at: { type: string, format: date-time }
        score: { $ref: "#/components/schemas/Score" }
    ListResult:
      type: object
      properties:
        runs:
          type: array
          items: { $ref: "#/components/schemas/RunSummary" }
        limit: { type: integer }
        next_cursor:
          type: string
          format: date-time
          description: Pass as `before` to fetch the next page. Absent when has_more is false.
        has_more: { type: boolean }